muda o tipo para bigdecimal

+++ src/main/java/dominio/bemImovel/BemImovel.java
@@ -1,5 +1,7 @@
 package dominio.bemImovel;
 
+import java.math.BigDecimal;
+
 public interface BemImovel {
-    public Double getValor();
+    public BigDecimal getValor();
 }
